refactor: split user route helpers
